file creation

+++ 0x15-file_io/0-read_textfile.c
@@ -32,7 +32,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	{
 		return (0);
 	}
-	if (fd < letters)
+	if (fd < (int) letters)
 	{
 		letters = fd;
 	}
